+++ src/storage/src/event.rs
@@ -77,7 +77,7 @@ pub async fn insert(
     let resource_type: i32 = resource_type.into();
     let typ: i32 = typ.into();
     let info = info.map(|info| info.into_bytes());
-    let ts = common::timestamp_millis();
+    let ts = common::timestamp_millis() as i64;
     sqlx::query(
         "INSERT INTO events (resource_type, resource_name, typ, info, ts) VALUES (?, ?, ?, ?, ?)",
     )
@@ -208,7 +208,7 @@ pub async fn search(
 }
 
 pub async fn delete_expired(day: usize) -> Result<()> {
-    let ts = common::timestamp_millis();
+    let ts = common::timestamp_millis() as i64;
     sqlx::query("DELETE FROM events WHERE ts < ?")
         .bind(ts - (day as i64) * 24 * 60 * 60 * 1000)
         .execute(POOL.get().unwrap())
